+++ api/player.ts
@@ -1,4 +1,4 @@
-import querySupabase from "../apiSupabase/querySupabase";
+import querySupabase from "./querySupabase";
 
 const route = "Player";
