+++ src/App.jsx
@@ -1,8 +1,8 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom"; // Link
-import Splash from "../pages/splash";
-import TestAPI from "../pages/testAPI";
-import Onboarding from "../pages/onboarding";
-import Game from "../pages/game";
+import Splash from "./pages/splash";
+// import TestAPI from "./pages/testAPI";
+import Onboarding from "./pages/onboarding";
+import Game from "./pages/game";
 
 import "./App.css";
 
@@ -12,7 +12,7 @@ function App() {
       <Router>
         <Routes>
           <Route path="/" element={<Splash />} />
-          <Route path="/testAPI" element={<TestAPI />} />
+          {/* <Route path="/testAPI" element={<TestAPI />} /> */}
           <Route path="/onboarding" element={<Onboarding />} />
           <Route path="/game" element={<Game />} />
         </Routes>
